Create MusicalFlags.py
This mod is a version of the classic game musical chairs.  However, instead of chairs there are flags which players must rush to grab to survive the round.  Points are awarded based on how many rounds a player survives, and the number of rounds is based on the number of players in the session.  Of course, there will be one winner at the final round at the end of each game.  Enjoy!